bot.py
import discord
import openai
import os
from dotenv import load_dotenv

# Secrets
load_dotenv()
openai.api_key = os.getenv("OPENAI_API_KEY")
discord_token = os.getenv("DISCORD_TOKEN")

# Discord client
intents = discord.Intents.default()
intents.message_content = True
client = discord.Client(intents=intents)

@client.event
async def on_ready():
    print(f'Bot conectado como {client.user}')

@client.event
async def on_message(message):
    if message.author == client.user:
        return

    if message.content.startswith("!senku "):
        prompt = message.content[len("!senku "):]

        response = openai.ChatCompletion.create(
            model="gpt-3.5-turbo",
            messages=[{"role": "user", "content": prompt}]
        )

        reply = response["choices"][0]["message"]["content"]
        await message.channel.send(reply)

# Cmon botsito
client.run(discord_token)
